refactor: add lamda layer and remake lamdas creation
