+++ package.json
@@ -11,8 +11,7 @@
     "test:coverage": "jest --coverage",
     "cdk": "cdk",
     "deploy": "cdk deploy",
-    "destroy": "cdk destroy",
-    "build:deploy": "npm run build && npm run deploy"
+    "destroy": "cdk destroy"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.138",
