+++ products-service/getProductsList.ts
@@ -1,11 +1,8 @@
-import { products } from './products';
-import prepareResponse from './prepareResponse';
+import { getProductsList, prepareResponse } from '/opt/utils';
 import type { HttpResponse } from '@/types';
 
-export const getProductsData = () => products;
-
 export const handler = async (): HttpResponse => {
-  const data = getProductsData();
+  const data = getProductsList();
 
   if (data.length) {
     return prepareResponse(200, data);
