+++ test/products-service/getProductById.test.ts
@@ -1,4 +1,4 @@
-import { handler } from "@/products/getProductById";
+import { handler } from "@/products-service/getProductById";
 import { products } from "../mock/products";
 import { httpEventMock } from "../mock/httpEventMock";
 import type { HttpEventRequest, ProductPathParams } from "@/types";
