+++ test/products-service/getProductsList.test.ts
@@ -1,4 +1,5 @@
-import * as getProducts from "@/products/getProducts";
+import * as utils from "/opt/utils";
+import { handler } from "@/products-service/getProductsList";
 import { products } from "../mock/products";
 
 beforeEach(() => {
@@ -7,13 +8,13 @@ beforeEach(() => {
 
 describe('getProducts handler', () => {
   it("Should return products array", async () => {
-    const response = await getProducts.handler();
+    const response = await handler();
     expect(JSON.parse(response.body)).toStrictEqual(products);
   });
 
   it("Should return statusCode 404", async () => {
-    jest.spyOn(getProducts, 'getProductsData').mockReturnValue([]);
-    const response = await getProducts.handler();
+    jest.spyOn(utils, 'getProductsList').mockReturnValue([]);
+    const response = await handler();
     expect(response.statusCode).toEqual(404);
   });
 })
